feat: add check for timer, return additional infor for main endpoints

+++ backend/app/db/crud/dislike_crud.py
@@ -1,9 +1,9 @@
 from fastapi import HTTPException
 from sqlalchemy.orm import Session
 from starlette import status
-import typing as t
 
-from app.db import models, schemas
+from app.db import models
+from db.crud.book_crud import book_in_review_time
 
 
 def get_dislike(db: Session, user_id: int, review_id: int):
@@ -16,25 +16,17 @@ def get_dislike(db: Session, user_id: int, review_id: int):
     )
 
 
-def get_dislikes(
-    db: Session, skip: int = 0, limit: int = 100
-) -> t.List[schemas.DislikeOut]:
-    return db.query(models.Dislike).offset(skip).limit(limit).all()
-
-
-def create_dislike(db: Session, dislike: schemas.DislikeCreate):
-    user = db.query(models.User).filter(models.User.id == dislike.user_id).first()
-    if not user:
-        raise HTTPException(status_code=400, detail="User not found")
-
-    # Check if the review exists
-    review = (
-        db.query(models.Review).filter(models.Review.id == dislike.review_id).first()
-    )
+def create_dislike(db: Session, user_id: int, review_id: int):
+    review = db.query(models.Review).filter(models.Review.id == review_id).first()
     if not review:
         raise HTTPException(status_code=400, detail="Review not found")
 
-    db_dislike = models.Dislike(user_id=dislike.user_id, review_id=dislike.review_id)
+    if not book_in_review_time(review.book.created_at, review.book.review_hour_amount):
+        raise HTTPException(
+            status_code=400, detail="The book review is over, you cannot dislike it"
+        )
+
+    db_dislike = models.Dislike(user_id=user_id, review_id=review_id)
     db.add(db_dislike)
     db.commit()
     return db_dislike
@@ -44,6 +36,13 @@ def delete_dislike(db: Session, user_id: int, review_id: int):
     dislike = get_dislike(db, user_id, review_id)
     if not dislike:
         raise HTTPException(status.HTTP_404_NOT_FOUND, detail="Dislike not found")
+
+    if not book_in_review_time(
+        dislike.review.book.created_at, dislike.review.book.review_hour_amount
+    ):
+        raise HTTPException(
+            status_code=400, detail="The book review is over, you cannot dislike it"
+        )
     db.delete(dislike)
     db.commit()
     return dislike
